Attempts to model a `Cursor` with an `Event` where the read methods are pushed into it.
Implements very trivial binary and text format parsers to illustrate.

This implements the underlying readers through a `* mut` which is made safe
due to only accessing it through a mutable borrow in the `Event`.  GAT would
eliminate this, but it demonstrates how we can implement this.

Implements `PossibleCursor<B, T>` and generic example test.

Prototypes #180.