Breaks out ScalarType from ScalarValue
This introduces the `ScalarType` tag as distinct from a lazy
`ScalarValue` that may be behind a thunk (and thus deferred).
This allows us to "observe" the potential type of scalar without
having to evaluate it.

`ScalarThunk` is now a named tuple struct whose contents are
`pub(crate)` visible to avoid the sharp edge that the `ScalarType`
and `Thunk<'a, ScalarValue>` *must* align.  The net effect of this
is that external users cannot make lazy `ScalarThunk` instances,
but we can cross that bridge when we get there.