Changes the lifetime of TokenStream::next_token
Makes it return `AnnotatedToken<'a>` which means it cannot be tracked
statically.  Adds `UnderlyingReader<R>` to track this state and ensure
that the thunks returned by `next_token` invalidate themselves as to
not return invalid data.  This is required to support modeling an
`IonReader` over a `TokenStream` as the method that calls `next_token`
is not tied to when a `read_XXX` in any lexical context that allows us
to model the borrow at the point of `next_token`.

Fixes the lifetime of the stored token in `TokenStreamReader` to
explicitly operate on the same lifetime bounds as the enclosed
`TokenStream` with these API changes.

Adds more `memoize` methods to make the in-place materialization of an
event easier to use.

TODO: make `Token` easier to use from a `materialize` call.