Refactors ParamType
Renames `TaglessType` to `PrimitiveType`. Changes spellings
of `var_sym`/`var_str`/`var_int`/`var_uint` to have underscores.

Removes `ArgCardinality`, adding `ParameterMode` in its place to
capture `Normal`/`Rest` modes (not sure if this is better than
encapsulating it in an enum).

In `ParamType`, removes the value type and renames `val_cardinality`
to `stream_cardinality` to be more consistent.

Adds `StaticType::is_tagged` to query if a type is tagged/tagless.